refactor: 검증에 포함되어 있던 분기 MatchProgress로 분리
