+++ badminton-api/src/main/java/org/badminton/api/match/SinglesMatchProgress.java
@@ -1,12 +1,17 @@
 package org.badminton.api.match;
 
+import java.util.ArrayList;
 import java.util.List;
 
+import org.badminton.api.common.exception.match.MatchDuplicateException;
 import org.badminton.api.common.exception.match.MatchNotExistException;
 import org.badminton.api.match.model.dto.MatchDetailsResponse;
+import org.badminton.api.match.model.dto.MatchResponse;
 import org.badminton.api.match.model.dto.SetScoreUpdateRequest;
 import org.badminton.api.match.model.dto.SetScoreUpdateResponse;
 import org.badminton.domain.common.enums.MatchType;
+import org.badminton.domain.league.entity.LeagueEntity;
+import org.badminton.domain.league.entity.LeagueParticipantEntity;
 import org.badminton.domain.match.model.entity.SinglesMatchEntity;
 import org.badminton.domain.match.model.entity.SinglesSetEntity;
 import org.badminton.domain.match.repository.SinglesMatchRepository;
@@ -17,6 +22,15 @@ import lombok.AllArgsConstructor;
 public class SinglesMatchProgress implements MatchProgress {
 
 	private SinglesMatchRepository singlesMatchRepository;
+
+	@Override
+	public List<MatchResponse> makeMatches(LeagueEntity league, List<LeagueParticipantEntity> leagueParticipantList) {
+		List<SinglesMatchEntity> singlesMatches = makeSinglesMatches(leagueParticipantList, league);
+		return singlesMatches
+			.stream()
+			.map(MatchResponse::entityToSinglesMatchResponse)
+			.toList();
+	}
 
 	@Override
 	public List<MatchDetailsResponse> initDetails(Long leagueId) {
@@ -39,9 +53,25 @@ public class SinglesMatchProgress implements MatchProgress {
 		singlesMatch.getSinglesSets()
 			.get(setIndex - 1)
 			.saveSetScore(setScoreUpdateRequest.score1(), setScoreUpdateRequest.score2());
+
+		// 승자에 따라 Match에 이긴 세트수를 업데이트해준다. 만약 2번을 모두 이긴 참가자가 있다면 해당 Match는 종료된다.
+		if (setScoreUpdateRequest.score1() > setScoreUpdateRequest.score2()) {
+			singlesMatch.player1WinSet();
+		} else
+			singlesMatch.player2WinSet();
+
 		singlesMatchRepository.save(singlesMatch);
 		return SetScoreUpdateResponse.singlesSetentityToSetScoreUpdateResponse(
 			singlesMatch.getSinglesSets().get(setIndex - 1));
+	}
+
+	@Override
+	public void checkDuplicateMatchInLeague(Long leagueId, MatchType matchType) {
+		singlesMatchRepository.findByLeague_LeagueId(leagueId).ifPresent(
+			singlesMatch -> {
+				throw new MatchDuplicateException(matchType, singlesMatch.getSinglesMatchId());
+			}
+		);
 	}
 
 	private SinglesMatchEntity initSinglesMatch(SinglesMatchEntity singlesMatch) {
@@ -56,6 +86,20 @@ public class SinglesMatchProgress implements MatchProgress {
 
 		singlesMatchRepository.save(singlesMatch);
 		return singlesMatch;
+	}
+
+	private List<SinglesMatchEntity> makeSinglesMatches(List<LeagueParticipantEntity> leagueParticipantList,
+		LeagueEntity league) {
+
+		List<SinglesMatchEntity> singlesMatches = new ArrayList<>();
+		for (int i = 0; i < leagueParticipantList.size() - 1; i += 2) {
+			SinglesMatchEntity singlesMatch = new SinglesMatchEntity(
+				league, leagueParticipantList.get(i), leagueParticipantList.get(i + 1)
+			);
+			singlesMatches.add(singlesMatch);
+			singlesMatchRepository.save(singlesMatch);
+		}
+		return singlesMatches;
 	}
 }
