+++ badminton-api/src/main/java/org/badminton/api/match/service/MatchProgressService.java
@@ -1,12 +1,9 @@
 package org.badminton.api.match.service;
 
-import java.util.List;
-
 import org.badminton.api.common.exception.league.LeagueNotExistException;
 import org.badminton.api.match.DoublesMatchProgress;
 import org.badminton.api.match.MatchProgress;
 import org.badminton.api.match.SinglesMatchProgress;
-import org.badminton.api.match.model.dto.MatchDetailsResponse;
 import org.badminton.api.match.model.dto.SetScoreUpdateRequest;
 import org.badminton.api.match.model.dto.SetScoreUpdateResponse;
 import org.badminton.domain.common.enums.MatchType;
@@ -25,19 +22,6 @@ public class MatchProgressService {
 	private final DoublesMatchRepository doublesMatchRepository;
 
 	private final LeagueRepository leagueRepository;
-
-	public List<MatchDetailsResponse> initMatchDetails(Long clubId, Long leagueId) {
-		// 경기 일정이 있는지 확인하고 꺼내기
-		LeagueEntity league = leagueRepository.findById(leagueId)
-			.orElseThrow(() -> new LeagueNotExistException(clubId, leagueId));
-
-		// 경기 타입을 구분하여 처리하기 위해
-		MatchType matchType = league.getMatchType();
-
-		// 경기 타입을 구분하여 처리하기 위해
-		MatchProgress matchProgress = createMatchProgress(matchType);
-		return matchProgress.initDetails(clubId);
-	}
 
 	// 한 세트가 끝나고 세트 결과를 저장
 	public SetScoreUpdateResponse updateSetScore(Long clubId, Long leagueId, Long matchId, int setIndex,
